Defining standard sizes for images
Previously, I had used inline styles on the image placeholder div. This
won't work because they get passed along to the RSS feed content
_without_ the rest of the styles needed for them to make sense.

From an art direction standpoint, it shouldn't be too bad to stick to
standard image ratios 1:1, 4:3, 3:2, 16:9. I added a CSS class for each
ratio in both portrait and landscape orientation.

I have to go back and adjust the existing images and placeholders now,
but it's better to do it sooner than later.